+++ src/test/java/BaseTest/Hooks.java
@@ -1,11 +1,17 @@
 package BaseTest;
 
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.testng.annotations.*;
 import org.testng.asserts.SoftAssert;
 import utils.Driver;
 import utils.Pages;
 
+
+
 public class Hooks {
+	private static Logger logger = LogManager.getLogger(Hooks.class);
 
 	protected static Pages pages;
 	protected SoftAssert softAssert = new SoftAssert();
@@ -13,8 +19,12 @@ public class Hooks {
 	@Parameters("browserType")
 	public static void setUpTestEnvironment(@Optional("chrome") String browserType) {
 		Driver.setDriver(browserType);
+
+		logger.info("Starting Tests");
+		logger.info("Browser Type:" + browserType);
 		pages = new Pages();
 		pages.getHomePage().clickOnWebAutomationLink();
+		logger.debug("Web Automation link clicked");
 		if (browserType.equalsIgnoreCase("firefox")) {
 			try {
 				Thread.sleep(2000);
@@ -27,6 +37,7 @@ public class Hooks {
 	@AfterTest
 	public static void tearDown() {
 		Driver.closeDriver();
+		logger.info("Finished Tests");
 	}
 
 	public void waitSec(int sec) {