+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # CHANGE LOG
 
-## [1.0.4]
+## [1.0.5]
 
 ### Changed
+- Log4J Implementation
 - [Assignment_18 Part 2 is completed](https://inaracademygroup3.atlassian.net/browse/SCRUM-101?atlOrigin=eyJpIjoiY2VjMDcxYzY1NjUxNDY0ZjljYzYyYzYwZWZmNDA2NDEiLCJwIjoiaiJ9)
 - TestNG Framework Implementation
 - [Assignment_18 Part 1 is completed](https://inaracademygroup3.atlassian.net/browse/SCRUM-83?atlOrigin=eyJpIjoiZjU0ZjliZDVhNGVhNDg5NWE3NWJkMWJjMDE3ODM4NWQiLCJwIjoiaiJ9)
@@ -10,4 +11,4 @@
 
 [groupId]: org.inar
 [artifactId]: inar.webAutomation
-[version]: 1.0.4
+[version]: 1.0.5